use_sidecar_extension/fastapiapp/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
httpx>=0.25.0
python-dotenv>=1.0.0
sse-starlette>=1.6.5
pydantic>=2.4.2
jinja2>=3.1.2